Fixes according to PR review
